Cleaning of max activation analysis code
