+++ channel_vis.py
@@ -1,12 +1,5 @@
-'''
-Filter visualization techniques derived from those discussed in
-'How to Visualize Convolutional Features in 40 Lines of Code', by Fabio M. Graetz.
-
-https://towardsdatascience.com/how-to-visualize-convolutional-features-in-40-lines-of-code-70b7d87b0030
-'''
 import torch
 from torchvision import datasets, models
-from torch.autograd import Variable
 import numpy as np
 import cv2
 import matplotlib.pyplot as plt
@@ -14,26 +7,20 @@ import argparse
 import sys
 import os
 
+from utils.activation_analysis import LayerActivationAnalysis
 from models.UNet import UNet
 
 # Maps an encoder layer index to the index of the child module associated with
 # that layer in the VGG11 model
 vgg11_layer_dict = {1:0, 2:3, 3:6, 4:8, 5:11, 6:13, 7:16, 8:18}
 
-def get_layer(model, layer_index):
+def get_conv_layer(model, layer_index):
     if isinstance(model, UNet):
         layer = model.get_encoder_layer(layer_index)
     elif isinstance(model, models.vgg.VGG):
         child_index = vgg11_layer_dict[layer_index]
         layer = list(model.children())[0][child_index]
     return layer
-
-def upscale_image(image, upscale_size):
-    # Upscales an image with dimensions (3,h,w) to (3,upscale_size,upscale_size)
-    image = np.transpose(image, axes=(1,2,0))
-    image = cv2.resize(image, (upscale_size,upscale_size), interpolation = cv2.INTER_CUBIC)
-    image = np.transpose(image, axes=(2,0,1))
-    return image
 
 def save_image(image_array, destination):
     plt.figure()
@@ -54,77 +41,7 @@ def save_image_grid(image_arrays, destination, width=16, height=16):
     plt.savefig(destination, transparent=True, bbox_inches='tight', pad_inches=0)
     plt.close()
 
-
-class LayerActivations():
-    def __init__(self, layer):
-        self.forward_hook = layer.register_forward_hook(self.hook_fn)
-
-    def hook_fn(self, module, input, output):
-        self.activations = output
-
-    def remove_hook(self):
-        self.forward_hook.remove()
-
-
-class LayerActivationAnalysis():
-    def __init__(self, model, layer):
-        self.model = model
-        self.layer = layer
-
-    def set_layer(self, layer):
-        self.layer = layer
-
-    def get_activated_filter_indices(self, initial_img_size=56):
-        layer_activations = LayerActivations(self.layer)
-        image = (np.random.uniform(0, 255, size=(3,initial_img_size,initial_img_size)) / 255).astype(np.float32, copy=False)
-        image_tensor = torch.from_numpy(image).expand(1, -1, -1, -1)
-
-        _ = self.model(image_tensor)
-
-        filter_activations = layer_activations.activations[0].detach().numpy()
-
-        layer_activations.remove_hook()
-
-        return np.unique(np.nonzero(filter_activations)[0])
-
-    def get_max_activating_image(self, filter_index, initial_img_size=56, upscaling_steps=12, upscaling_factor=1.2, lr=0.01, update_steps=15, verbose=False):
-        device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
-        self.model.to(device)
-
-        layer_activations = LayerActivations(self.layer)
-        size = initial_img_size
-        image = (np.random.uniform(0, 255, size=(3,size,size)) / 255).astype(np.float32, copy=False)
-        for i in range(upscaling_steps):
-            image_tensor = torch.from_numpy(image).expand(1, -1, -1, -1)
-            image_tensor = Variable(image_tensor, requires_grad=True)
-            image_tensor = image_tensor.to(device)
-
-            if not image_tensor.grad is None:
-                image_tensor.grad.zero_()
-
-            optimizer = torch.optim.Adam([image_tensor], lr=lr, weight_decay=1e-6)
-
-            # Update image update_steps times
-            for n in range(update_steps):
-                optimizer.zero_grad()
-                _ = self.model(image_tensor)
-                loss = -1 * (layer_activations.activations[0, filter_index].norm())
-                if verbose and (n % 5 == 0):
-                    print('Loss at upscale step {}/{}, update {}/{}: {}'
-                            .format(i, upscaling_steps-1, n, update_steps-1, loss))
-                loss.backward()
-                optimizer.step()
-
-            image = torch.squeeze(image_tensor, dim=0).clone().detach().numpy()
-            size = int(upscaling_factor * size)
-            image = upscale_image(image, size)
-
-        output = np.transpose(image, axes=(1,2,0))
-        layer_activations.remove_hook()
-        return np.clip(output, 0, 1)
-
-
-if __name__=='__main__':
+def get_cli_arguments():
     parser = argparse.ArgumentParser(description='Analyze encoder module filters of a trained model')
 
     parser.add_argument('--unet',
@@ -151,6 +68,31 @@ if __name__=='__main__':
                         nargs='*',
                         help='Layer output channels to visualize')
 
+    parser.add_argument('--img_size',
+                        type=int,
+                        default=56,
+                        help='Initial size of the randomly initialized image')
+
+    parser.add_argument('--upscale_steps',
+                        type=int,
+                        default=12,
+                        help='Number of upscaling steps while optimizing the maximally activating image')
+
+    parser.add_argument('--upscale_factor',
+                        type=float,
+                        default=1.2,
+                        help='Upscaling factor at each upscaling step')
+
+    parser.add_argument('--lr',
+                        type=float,
+                        default=0.01,
+                        help='Learning rate for image optimization update step')
+
+    parser.add_argument('--steps',
+                        type=int,
+                        default=15,
+                        help='Number of optimization update steps at each upscaling step')
+
     parser.add_argument('--grid',
                         action='store_true',
                         help='Visualize multiple output channels as a grid')
@@ -163,7 +105,11 @@ if __name__=='__main__':
     parser.add_argument('-v', '--verbose',
                         action='store_true',
                         help='Output status of image optimization processes')
-    args = parser.parse_args()
+    return parser.parse_args()
+
+
+if __name__=='__main__':
+    args = get_cli_arguments()
 
     if args.unet:
         if not os.path.exists(args.checkpoint):
@@ -179,14 +125,17 @@ if __name__=='__main__':
     else:
         sys.exit('No model provided, please specify --unet or --vgg11 to analyze the UNet or VGG11 encoder, respectively')
 
-
+    # Set model to evaluation mode and fix the parameter values
     model.eval()
     for param in model.parameters():
         param.requires_grad_(False)
-    layer = get_layer(model, args.layer)
+
+    layer = get_conv_layer(model, args.layer)
 
     analyzer = LayerActivationAnalysis(model, layer)
 
+
+    # Save a grid of channel activation visualizations
     if args.grid:
         if args.channels is not None:
             channels = args.channels
@@ -200,21 +149,38 @@ if __name__=='__main__':
         for i, channel in enumerate(channels):
             if args.verbose:
                 print('Generating image {} of {}...'.format(i+1, len(channels)))
-            img = analyzer.get_max_activating_image(channel, verbose=args.verbose)
+
+            img = analyzer.get_max_activating_image(channel,
+                                                    initial_img_size=args.img_size,
+                                                    upscaling_steps=args.upscale_steps,
+                                                    upscaling_factor=args.upscale_factor,
+                                                    lr=args.lr,
+                                                    update_steps=args.steps,
+                                                    verbose=args.verbose)
+
             imgs.append(img)
         channel_string = '-'.join(str(channel_id) for channel_id in channels)
         output_dest = os.path.join(args.out, '{}_layer{}_channels{}.png'.format(model_name, args.layer, channel_string))
         save_image_grid(imgs, output_dest)
 
+    # Save a channel activation visualization for each specified channel
     elif args.channels is not None:
         for channel in args.channels:
-            img = analyzer.get_max_activating_image(channel, verbose=args.verbose)
+
+            img = analyzer.get_max_activating_image(channel,
+                                                    initial_img_size=args.img_size,
+                                                    upscaling_steps=args.upscale_steps,
+                                                    upscaling_factor=args.upscale_factor,
+                                                    lr=args.lr,
+                                                    update_steps=args.steps,
+                                                    verbose=args.verbose)
 
             output_dest = os.path.join(args.out, '{}_layer{}_channel{}.png'.format(model_name, args.layer, channel))
             save_image(img, output_dest)
 
+    # Output the channels activated by a randomly initialize image
     else:
-        activated_channels = analyzer.get_activated_filter_indices()
+        activated_channels = analyzer.get_activated_filter_indices(initial_img_size=args.img_size)
         print('Output channels in conv layer {} activated by random image input:'.format(args.layer))
         print(activated_channels)
         print()
